zerowaste-ai/ai-service/requirements.txt
flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.5.0
pandas==2.2.2
numpy==1.26.4
joblib==1.4.2
python-dotenv==1.0.1
gunicorn==22.0.0
